player health UI added, health regen, display and damage

+++ Assets/Scripts/Spike.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Spike : MonoBehaviour
 {
-    public int damageAmount = 1;
+    public float damageAmount = 1f;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
